requirements.txt
alembic==1.7.7
async-timeout==4.0.2
Babel==2.9.1
bcrypt==3.2.0
bleach==4.1.0
blinker==1.4
Bootstrap-Flask==2.0.2
cffi==1.15.0
chardet==4.0.0
click==8.0.4
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.1.3
Flask==2.0.3
Flask-Babel==2.0.0
Flask-BabelEx==0.9.4
Flask-Excel==0.0.7
Flask-Executor==0.10.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-Reuploaded==1.2.0
Flask-Security-Too==4.1.3
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==1.0.0
gurobipy==10.0.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
lml==0.1.0
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.1.1
Mosek==10.0.27
numpy==1.23.4
packaging==21.3
pandas==1.4.1
passlib==1.7.4
pycparser==2.21
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
scipy==1.8.0
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.32
SQLAlchemy-Utils==0.38.2
texttable==1.6.4
webencodings==0.5.1
Werkzeug==2.0.3
wrapt==1.14.0
WTForms==3.0.1
